refactor: update linkedIdType filename
