+++ libs/common/src/misc/linkedFieldOption.decorator.ts
@@ -1,4 +1,4 @@
-import { LinkedIdType } from "../enums/linkedIdType";
+import { LinkedIdType } from "../enums/linked-id-type";
 import { ItemView } from "../vault/models/view/item.view";
 
 export class LinkedMetadata {
