+++ libs/common/src/models/export/field.export.ts
@@ -1,5 +1,5 @@
 import { FieldType } from "../../enums/field-type";
-import { LinkedIdType } from "../../enums/linkedIdType";
+import { LinkedIdType } from "../../enums/linked-id-type";
 import { Field as FieldDomain } from "../../vault/models/domain/field";
 import { FieldView } from "../../vault/models/view/field.view";
 import { EncString } from "../domain/enc-string";
