+++ libs/common/src/vault/models/data/field.data.ts
@@ -1,5 +1,5 @@
 import { FieldType } from "../../../enums/field-type";
-import { LinkedIdType } from "../../../enums/linkedIdType";
+import { LinkedIdType } from "../../../enums/linked-id-type";
 import { FieldApi } from "../../../models/api/field.api";
 
 export class FieldData {
