+++ libs/common/src/vault/models/view/card.view.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { CardLinkedId as LinkedId } from "../../../enums/linkedIdType";
+import { CardLinkedId as LinkedId } from "../../../enums/linked-id-type";
 import { linkedFieldOption } from "../../../misc/linkedFieldOption.decorator";
 
 import { ItemView } from "./item.view";
