+++ libs/common/src/vault/models/view/field.view.ts
@@ -1,7 +1,7 @@
 import { Jsonify } from "type-fest";
 
 import { FieldType } from "../../../enums/field-type";
-import { LinkedIdType } from "../../../enums/linkedIdType";
+import { LinkedIdType } from "../../../enums/linked-id-type";
 import { View } from "../../../models/view/view";
 import { Field } from "../domain/field";
